+++ src/main/java/com/partnerd/repository/projectRepository/projectComment/ProjectCommentRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.partnerd.repository.projectRepository;
+package com.partnerd.repository.projectRepository.projectComment;
 
 import com.partnerd.domain.ProjectComment;
 import com.partnerd.domain.QMember;
