+++ src/main/java/com/partnerd/repository/projectRepository/projectVote/ProjectVoteRepositoryCustom.java
@@ -1,4 +1,4 @@
-package com.partnerd.repository.projectRepository;
+package com.partnerd.repository.projectRepository.projectVote;
 
 
 public interface ProjectVoteRepositoryCustom {
