+++ src/main/java/com/partnerd/repository/projectRepository/promotionProjectComment/PromotionProjectCommentRepositoryCustom.java
@@ -1,4 +1,4 @@
-package com.partnerd.repository.projectRepository;
+package com.partnerd.repository.projectRepository.promotionProjectComment;
 
 
 import com.partnerd.domain.PromotionProjectComment;
