+++ src/main/java/com/partnerd/service/homeService/impl/HomeServiceImpl.java
@@ -2,8 +2,8 @@ package com.partnerd.service.homeService.impl;
 
 import com.partnerd.repository.clubRepository.ClubRepository;
 import com.partnerd.repository.collabPostRepository.collabPost.CollabPostRepository;
-import com.partnerd.repository.projectRepository.ProjectRepository;
-import com.partnerd.repository.projectRepository.PromotionProjectRepository;
+import com.partnerd.repository.projectRepository.project.ProjectRepository;
+import com.partnerd.repository.projectRepository.promotionProject.PromotionProjectRepository;
 import com.partnerd.service.homeService.HomeService;
 import com.partnerd.web.dto.homeDTO.response.HomeClubDTO;
 import com.partnerd.web.dto.homeDTO.response.HomeCollabPostDTO;
