+++ src/main/java/com/partnerd/service/projectService/ProjectCommentServiceImpl.java
@@ -7,9 +7,9 @@ import com.partnerd.domain.Member;
 import com.partnerd.domain.ProjectComment;
 import com.partnerd.domain.mapping.ProjectLikes;
 import com.partnerd.repository.memberRepository.MemberRepository;
-import com.partnerd.repository.projectRepository.ProjectCommentRepository;
-import com.partnerd.repository.projectRepository.ProjectLikesRepository;
-import com.partnerd.repository.projectRepository.ProjectRepository;
+import com.partnerd.repository.projectRepository.projectComment.ProjectCommentRepository;
+import com.partnerd.repository.projectRepository.projectLikes.ProjectLikesRepository;
+import com.partnerd.repository.projectRepository.project.ProjectRepository;
 import com.partnerd.web.dto.projectDTO.ProjectCommentRequestDTO;
 import com.partnerd.web.dto.projectDTO.ProjectCommentResponseDTO;
 import lombok.RequiredArgsConstructor;
