+++ src/main/java/com/partnerd/service/projectService/ProjectServiceImpl.java
@@ -13,19 +13,13 @@ import com.partnerd.repository.memberRepository.MemberRepository;
 import com.partnerd.repository.projectRepository.ProjectCategoryPreferRepository;
 import com.partnerd.repository.projectRepository.ProjectCategoryRepository;
 import com.partnerd.repository.projectRepository.ProjectMemberRepository;
-import com.partnerd.repository.projectRepository.ProjectRepository;
+import com.partnerd.repository.projectRepository.project.ProjectRepository;
 import com.partnerd.web.dto.projectDTO.ProjectRequestDTO;
-import com.querydsl.jpa.JPQLQuery;
-import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Date;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
