+++ src/main/java/com/partnerd/service/projectService/PromotionProjectCommentServiceImpl.java
@@ -7,9 +7,9 @@ import com.partnerd.domain.Member;
 import com.partnerd.domain.PromotionProjectComment;
 import com.partnerd.domain.mapping.ProjectLikes;
 import com.partnerd.repository.memberRepository.MemberRepository;
-import com.partnerd.repository.projectRepository.ProjectLikesRepository;
-import com.partnerd.repository.projectRepository.PromotionProjectCommentRepository;
-import com.partnerd.repository.projectRepository.PromotionProjectRepository;
+import com.partnerd.repository.projectRepository.projectLikes.ProjectLikesRepository;
+import com.partnerd.repository.projectRepository.promotionProjectComment.PromotionProjectCommentRepository;
+import com.partnerd.repository.projectRepository.promotionProject.PromotionProjectRepository;
 import com.partnerd.web.dto.projectDTO.PromotionProjectCommentRequestDTO;
 import com.partnerd.web.dto.projectDTO.PromotionProjectCommentResponseDTO;
 import lombok.RequiredArgsConstructor;
