+++ src/main/java/com/partnerd/service/projectService/PromotionProjectServiceImpl.java
@@ -8,10 +8,10 @@ import com.partnerd.domain.*;
 import com.partnerd.domain.enums.ImageType;
 import com.partnerd.domain.mapping.ProjectVote;
 import com.partnerd.domain.mapping.PromotionProjectMember;
-import com.partnerd.repository.projectRepository.ProjectVoteRepository;
+import com.partnerd.repository.projectRepository.projectVote.ProjectVoteRepository;
 import com.partnerd.repository.projectRepository.PromotionProjectMemberRepository;
 import com.partnerd.repository.memberRepository.MemberRepository;
-import com.partnerd.repository.projectRepository.PromotionProjectRepository;
+import com.partnerd.repository.projectRepository.promotionProject.PromotionProjectRepository;
 import com.partnerd.web.dto.projectDTO.PromotionProjectRequestDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
